style: 优化代码结构，分离为 WebAPI 逻辑与实际请求逻辑。
